tema4/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
azure-cosmos
azure-storage
pytz
sendgrid
O365
azure-keyvault-secrets
azure-keyvault-keys
azure-keyvault-certificates
azure-identity
msrestazure